requirements.txt
aiohttp
pillow
aiogram
yt-dlp
python-youtube
cairosvg
thefuzz
